+++ App/frontend/package.json
@@ -23,7 +23,6 @@
     "dotenv": "^16.0.0",
     "ejs": "^3.1.6",
     "express": "^4.17.2",
-    "express-flash": "^0.0.2",
     "express-session": "^1.17.2",
     "mysql2": "^2.3.3",
     "sequelize": "^6.15.0"
